moved cloneutility class to new folder/namespace
