+++ Reveal.Sdk.Dom/Visualizations/Visualization.cs
@@ -1,5 +1,5 @@
 ﻿using Newtonsoft.Json;
-using Reveal.Sdk.Dom.Core;
+using Reveal.Sdk.Dom.Core.Utilities;
 using Reveal.Sdk.Dom.Data;
 using Reveal.Sdk.Dom.Filters;
 using Reveal.Sdk.Dom.Serialization.Converters;
